rename evaluation-types
